videodecoder: request sync-points regularly on error
If we are not receiving a sync-point for a very long time, we need to
keep asking for them. The request-sync-point logic keeps track of how
many keyunitrequests we are allowed to send, but that would not matter
if we don't keep asking.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/930>