ges:uri-asset: Ignore streams with no ID
Since 0d95d9258b97 we respect the asset stream-id in `GESUriSource` so
we can not work with unknown or broken stream ID in the assets.

We just ignore them, warning about it and we should fix that in
demuxer so they don't expose pad without providing a stream id for them.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/1001>